refactor iframe page
