+++ src/page/index.ts
@@ -9,4 +9,4 @@ export { PaymentStepPage } from './payment-step.page';
 export { BankPaymentPage } from './bank-payment.page';
 export { OrderResumePage } from './order-resume.page';
 export { PersonalInformationPage } from './personal-information.page';
-export { IFramePage } from './IFrame.page';
+export { IFramePage } from './i-frame.page';
